+++ src/main/java/com/navangs/maribong/repository/PostLikeRepository.java
@@ -6,5 +6,5 @@ import java.util.List;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface PostLikeRepository extends JpaRepository<PostLike, PostLikeId> {
-    List<PostLike> findById_PostId(Integer postId);
+    List<PostLike> findById_PostId(Long postId);
 }
