+++ src/main/java/com/navangs/maribong/repository/PostPhotoRepository.java
@@ -5,5 +5,5 @@ import java.util.List;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface PostPhotoRepository extends JpaRepository<PostPhoto, Integer> {
-    List<PostPhoto> findByPostId(Integer postId);
+    List<PostPhoto> findByPostId(Long postId);
 }
