+++ src/main/java/com/navangs/maribong/repository/QuestionRepository.java
@@ -5,5 +5,5 @@ import java.util.List;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface QuestionRepository extends JpaRepository<Question, Integer> {
-    List<Question> findByQuizIdIn(List<Integer> quizIds);
+    List<Question> findByQuizIdIn(List<Long> quizIds);
 }
