+++ src/main/java/com/navangs/maribong/repository/QuizRepository.java
@@ -4,5 +4,5 @@ import com.navangs.maribong.domain.Quiz;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface QuizRepository extends JpaRepository<Quiz, Integer> {
-    Quiz findQuizById(Integer id);
+    Quiz findQuizById(Long id);
 }
